// frontend/src/types/illnessAssessment.ts
export interface IllnessAssessment {
  id: number;
  illness_group: string;
  illness_type: string;
  injury_detail: string;
  sickness: number;
  injury: number;
  burn: number;
  frostbite: number;
  other_diagnosis: number;
  diagnosis_worsening_condition: number;
  diagnosis_acute_coronary_syndrome: number;
  diagnosis_fainting: number;
  diagnosis_seizure: number;
  diagnosis_high_blood_pressure: number;
  diagnosis_intoxicated: number;
  diagnosis_stroke: number;
  diagnosis_copd: number;
  diagnosis_illness: number;
  diagnosis_infection: number;
  diagnosis_circulatory_dysregulation: number;
  diagnosis_psychological_issue: number;
  main_diagnosis_yes_or_no: number;
  diagnosis_heart_issues: number;
  diagnosis_heart_failure: number;
  diagnosis_injury: number;
  diagnosis_respiratory_disease: number;
  diagnosis_nervous_system_disease: number;
  diagnosis_fracture: number;
  diagnosis_pain: number;
  diagnosis_severity: number;
  infections: number;
  lung_disease: number;
  cardiovascular_disease: number;
  abdominal_disease: number;
  metabolic_disease: number;
  other_diseases: number;
  cns_disease: number;
  respiratory_disease: number;
  psychiatric_disease: number;
  gyn_obstetrics_emergencies: number;
}